Rename the file api-types
