+++ packages/sdk/src/index.ts
@@ -38,7 +38,7 @@ export * from './services/share-service';
 export * from './services/template-service';
 export * from './services/user-service';
 export * from './util/api-factory';
-export * from './util/api-model';
+export * from './util/api-types';
 export * from './util/codecs';
 export * from './util/constants';
 export * from './util/did-management';
